Addressed Copilot comments and reorganized some unit tests
