+++ projects/cuid/tests/CMakeLists.txt
@@ -31,6 +31,7 @@ enable_testing()
 add_executable(
     amdcuid_test
     unit/cuid_api_test.cc
+    unit/cuid_misc_test.cc
     functional/cuid_reverse_lookup_test.cc
 )
